refactor: move model to adapter repo
