+++ internal/adapter/adapter.go
@@ -3,11 +3,11 @@ package model
 import (
 	"database/sql"
 
-	"github.com/eyo-chen/expense-tracker-go/internal/model/icon"
-	"github.com/eyo-chen/expense-tracker-go/internal/model/maincateg"
-	"github.com/eyo-chen/expense-tracker-go/internal/model/subcateg"
-	"github.com/eyo-chen/expense-tracker-go/internal/model/transaction"
-	"github.com/eyo-chen/expense-tracker-go/internal/model/user"
+	"github.com/eyo-chen/expense-tracker-go/internal/adapter/repository/icon"
+	"github.com/eyo-chen/expense-tracker-go/internal/adapter/repository/maincateg"
+	"github.com/eyo-chen/expense-tracker-go/internal/adapter/repository/subcateg"
+	"github.com/eyo-chen/expense-tracker-go/internal/adapter/repository/transaction"
+	"github.com/eyo-chen/expense-tracker-go/internal/adapter/repository/user"
 )
 
 type Model struct {
