+++ internal/adapter/interfaces/interfaces.go
