+++ internal/adapter/repository/icon/converter.go
