+++ internal/adapter/repository/icon/factory.go
