+++ internal/adapter/repository/icon/icon.go
